@@ -32,31 +32,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/localize": "^17.3.12",
-    "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/platform-server": "^17.3.12",
-    "@angular/pwa": "^17.3.10",
-    "@angular/router": "^17.3.12",
-    "@angular/service-worker": "^17.3.12",
-    "@angular/ssr": "^17.3.10",
+    "@angular/animations": "^18.2.13",
+    "@angular/cdk": "^18.2.13",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/localize": "^18.2.13",
+    "@angular/material": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/platform-server": "^18.2.13",
+    "@angular/pwa": "^18.2.12",
+    "@angular/router": "^18.2.13",
+    "@angular/service-worker": "^18.2.13",
+    "@angular/ssr": "^18.2.12",
     "@apollo/client": "^3.7.14",
     "@bhayward93/ckeditor5-build-minds": "^1.1.8",
+    "@bhayward93/web3modal-angular": "^18.2.13",
     "@ckeditor/ckeditor5-angular": "^5.2.0",
     "@ctrl/ngx-emoji-mart": "^9.2.0",
-    "@danielmoncada/angular-datetime-picker": "^17.0.0",
+    "@danielmoncada/angular-datetime-picker": "^18.0.0",
+    "@floating-ui/dom": "^1.6.12",
     "@fontsource/inter": "^4.5.15",
-    "@gorniv/ngx-universal": "^2.2.3",
-    "@mindsorg/web3modal-angular": "^17.3.12",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
-    "@popperjs/core": "^2.9.3",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.1",
     "@sentry/angular": "^8.26.0",
     "@sentry/node": "^8.26.0",
     "@tweenjs/tween.js": "^18.6.4",
@@ -171,7 +170,7 @@
     "storybook-addon-themes": "^6.1.0",
     "ts-node": "^10.9.1",
     "tslint": "~6.1.0",
-    "typescript": "5.2",
+    "typescript": "5.4",
     "webpack-bundle-analyzer": "^3.6.0",
     "webpack-cli": "^4.7.2",
     "xmlbuilder2": "^2.1.2"
