@@ -38,10 +38,12 @@ import { BlockchainMarketingModule } from './modules/blockchain/marketing/market
 import { BrandingModule } from './modules/branding/branding.module';
 import { CommentsModule } from './modules/comments/comments.module';
 import { JobsMarketingModule } from './modules/jobs/jobs.module';
-import { HttpClientModule } from '@angular/common/http';
+import {
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
 //import { ChannelContainerModule } from './modules/channel-container/channel-container.module';
 import { CodeHighlightModule } from './modules/code-highlight/code-highlight.module';
-import { CookieModule } from '@gorniv/ngx-universal';
 import { HomepageModule } from './modules/homepage/homepage.module';
 import { OnboardingV2Module } from './modules/onboarding-v2/onboarding.module';
 import { ConfigsService } from './common/services/configs.service';
@@ -66,15 +68,14 @@ import { MindsSentryErrorHandler } from './common/services/diagnostics/sentry-er
 @NgModule({
   bootstrap: [Minds],
   declarations: [Minds, Pages],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
   imports: [
-    CookieModule.forRoot(),
     ApolloModule,
     MarkdownModule.forRoot(),
     // TransferHttpCacheModule,
     BrowserAnimationsModule,
     ReactiveFormsModule,
     FormsModule,
-    HttpClientModule,
     CaptchaModule,
     LayoutModule,
     MindsFormsModule,
@@ -132,7 +133,7 @@ import { MindsSentryErrorHandler } from './common/services/diagnostics/sentry-er
       deps: [ConfigsService, Sentry.TraceService],
       multi: true,
     },
+    provideHttpClient(withInterceptorsFromDi()),
   ],
-  schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
 export class MindsModule {}
