@@ -1,9 +1,13 @@
 import { APP_BASE_HREF, CommonModule, Location } from '@angular/common';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import {
+  HttpClient,
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
 import { APP_ID, APP_INITIALIZER, NgModule, PLATFORM_ID } from '@angular/core';
 import { TransferState } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { CookieModule, CookieService } from '@gorniv/ngx-universal';
+import { CookieService } from '../../common/services/cookie.service';
 import { MindsHttpClient } from '../../common/api/client.service';
 import { BlockListService } from '../../common/services/block-list.service';
 import { ConfigsService } from '../../common/services/configs.service';
@@ -25,13 +29,8 @@ const routes = [{ path: 'embed/:guid', component: EmbeddedVideoComponent }];
 @NgModule({
   declarations: [EmbedComponent, EmbeddedVideoComponent],
   exports: [EmbedComponent],
-  imports: [
-    CommonModule,
-    HttpClientModule,
-    VideoModule,
-    CookieModule.forRoot(),
-    RouterModule.forRoot(routes, {}),
-  ],
+  bootstrap: [EmbedComponent],
+  imports: [CommonModule, VideoModule, RouterModule.forRoot(routes, {})],
   providers: [
     SiteService,
     {
@@ -75,7 +74,7 @@ const routes = [{ path: 'embed/:guid', component: EmbeddedVideoComponent }];
       deps: [Client, BlockListService],
     },
     { provide: APP_ID, useValue: 'm-app' },
+    provideHttpClient(withInterceptorsFromDi()),
   ],
-  bootstrap: [EmbedComponent],
 })
 export class EmbedModule {}
