@@ -17,15 +17,14 @@ import { sessionMock } from '../../../tests/session-mock.spec';
 import { ActivatedRoute, convertToParamMap, Router } from '@angular/router';
 import { BehaviorSubject } from 'rxjs';
 import { CookieService } from '../../common/services/cookie.service';
-import {
-  CookieOptionsProvider,
-  COOKIE_OPTIONS,
-  CookieModule,
-} from '@gorniv/ngx-universal';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { LivestreamService } from '../../modules/composer/services/livestream.service';
 import { ComposerBoostService } from './services/boost.service';
 import { PermissionIntentsService } from '../../common/services/permission-intents.service';
+import {
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
 
 describe('Composer', () => {
   let comp: ComposerComponent;
@@ -46,7 +45,6 @@ describe('Composer', () => {
           outputs: ['onPost'],
         }),
       ],
-      imports: [CookieModule, HttpClientTestingModule],
       providers: [
         {
           provide: ComposerModalService,
@@ -88,13 +86,11 @@ describe('Composer', () => {
         },
         CookieService,
         {
-          provide: COOKIE_OPTIONS,
-          useValue: CookieOptionsProvider,
-        },
-        {
           provide: LivestreamService,
           useValue: MockService(LivestreamService),
         },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
       ],
     }).compileComponents();
     service = TestBed.inject(LivestreamService);
